Move deno in preparation for adding new clients
