+++ .vscode/settings.json
@@ -10,5 +10,9 @@
     "pwa-node": "${workspaceFolder}/.vscode/mise-tools/node"
   },
   "deno.path": "${workspaceFolder}/.vscode/mise-tools/deno",
-  "mise.configureExtensionsUseSymLinks": true
+  "mise.configureExtensionsUseSymLinks": true,
+  "deno.enablePaths": [
+    "scripts",
+    "src/clients/deno"
+  ]
 }
