+++ src/clients/deno/examples/tldraw.ts
@@ -1,4 +1,4 @@
-import { createWebView } from "../src/lib.ts";
+import { createWebView } from "../main.ts";
 import * as esbuild from "https://deno.land/x/esbuild@v0.24.0/wasm.js";
 
 const tldrawApp = `
